fixed binary conversion to handle 64-bit

+++ NumberConverter.h
@@ -2,14 +2,15 @@
 #include <iostream>
 #include <vector>
 #include <algorithm>
+#include <math.h>
 
 #include "DEFINITIONS.h"
 
 class NumberConverter {
 public:
-	int d2b(int dec);
-	int b2d(int bin);
+	long long d2b(long long dec);			//  Przyjmuje liczbê decymaln¹ i zwraca liczbê w postaci binarnej
+	long long b2d(long long bin);			//  Przyjmuje liczbê binarn¹ i zwraca liczbê w postaci decymalnej
 	
-	std::vector<int> vd2b(int dec);
+	std::vector<int> vd2b(long long dec); //  Przyjmuje liczbê decymaln¹ i zwraca wektor liczby w postaci binarnej
 };
